Properly clip background of content view
This change allows users to use custom items that have colored background and also use a popup with rounded corners.

Before the changes, the popup wouldn't display rounded corners for the custom view because its background wasn't clipped and it was drawing outside of the area of rounded drawable. After the change, when drawing, the content is properly clipped to the area of the background drawable and thus rounded corners work correctly.