Release/4.0.0 (#64)
* Moved `super.bindItem` call to the end of `ItemViewHolder#bindItem`method so that it's possible to override values set by it in callbacks.
Updated `MaterialPopupMenuBuilder.SectionHolder.label` and `MaterialPopupMenuBuilder.ItemHolder.label` to be of type `CharSequence` rather than `String` to allow the use of `Spannables`.
* Added an option to set a fixed width of the popup menu.
This can be used e.g. so that the popup menu has the same width as the anchor view. There was a sample added for this use case.
* Updates minimum supported SDK version to API 19.
Added an option to set a fixed width of the popup menu.
This can be used e.g. so that the popup menu has the same width as the anchor view. There was a sample added for this use case.
To achieve this effect an option to override offset was also added.
* Replaced xxLeft and xxRight attributes with xxStart and xxEnd RTL equivalents.